requirements.txt
boto3==1.28.53
botocore==1.31.53
jmespath==1.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
s3transfer==0.6.2
six==1.16.0
urllib3==1.26.16
